Scoring
Improved description in the README. Score the raw body of the page and
not the plain text (so that invisible text such as URLs are part of
the score). Page sorting is such that with equal score pages that
start with a number get sorted descending (in the hopes of putting
newer pages at the top, if ISO dates are part of page names).