Test id parameter with subdirectory
This tests a construct like the following in view.html:

<form action="/edit/" method="GET">
  <label for="id">New page:</label>
  <input id="id" type="text" spellcheck="false" name="id"
         accesskey="g" value="{{.Dir}}/{{.Today}}" required>
  <button>Edit</button>
</form>

For a while I tried to handle mixed situations where both URL path and
query parameter have to be handled but I got strange 500 status codes.
Basically I wanted to support the following and was unable to write
code and a test for it to make that work:

<form action="/edit/{{.Dir}}" method="GET">
  <label for="id">New page:</label>
  <input id="id" type="text" spellcheck="false" name="id"
         accesskey="g" value="{{.Today}}" required>
  <button>Edit</button>
</form>

I kept getting a strange 500 code and duplicate headers according to
the tests. Very strange.