Reload templates on demand
There is now a global variable to hold the parsed templates, and a
fsnotify watcher that watches the directory for any changes to the
templates. If a template is written to, a go routine is started which
waits for a second and then checks whether any of the files files that
were written to haven't been written to in the last second. If so, it
is assumed that writing is finished and the template is reloaded.

Thus, if file is written to twice, two go routines start, the first
one doesn't find a file that's old enough but the second one does, and
so the template is reloaded.

Previously, it would reload all the templates for every request that
needs a template. And before that, it would only load the templates as
the wiki started up.