Get rid of fsnotify
When uploading a file via the drop handler, check whether the saved
file is a valid template and if so, reload the template. There is no
point in using fsnotify unless we do the same thing for the index.
That would be interesting, in fact. But it's also complicated and
gives me a headache.