More logging improvements
Errors in p.save aren't logged, so the saveHandler must log them.

Errors in p.notify are logged, so the saveHandler must not log them.
That's because in this situation, the errors are not directly related
to page saving (which is what the saveHandler does). However, the user
must still be made aware of the error, so a HTTP error page is shown.

When backup fails (since it now returns an error), save does not
proceed.

The helper functions in notify.go using c as the variable for the
changes page, the index page, or a hashtag page now all use p.