Makefile improvements
Introduce a PREFIX variable for the install target so one can override
this. Document this in the README.

Add rules for a build target and for a binary target that depends on
all the Go files so that these can be used as prerequisites for other
targets. The goal is to avoid unnecessary recompilations.

Designate all the non-file targets as phony targets.