Make some of the types private
Don't export types unless necessary. Mostly because that looks weird
in Go documentation: when the functions acting on the types are
private the types themselves might as well be private.

Consequently, only the types that are used to write templates are
exported! For those types, the documentation has been improved (mostly
for Feed and Item).

In order to avoid nameclashes when hiding the types that didn't need
exporting, new naming schema was used. As these types were contains
for maps plus a mutex, they are called "stores" controlling access to
the maps: Accounts → accountStore, Index → indexStore, Template →
templateStore, Watches → watchStore.