Add handling of relative links to missing command
Links such as [up](../index) are now handled correctly by joining it
to the directory of the page containing this link. A page foo/bar with
such a link no longer registers as linking to a missing page if the
index page exists. Note that [up](..) still registers as a link to a
missing page because the code does not take redirects from directories
to index pages into account.