Rewrite staticCmd to use a lot of go routines
Starting a go routine that walks the directory tree, adding files to a
tasks channel and then closing it; starting a bunch of worker go
routines that take files from the tasks channel until there is nothing
left to do; putting errors on a results channel and when they’re done,
putting a true value on a done channel; a watcher go routine that
checks the done channel and if every worker is done, close the results
channel; and the main program goes through all the values on the
results channel and sets up some short-circuiting in the case of
errors and otherwise it prints the counter. What a glorious spaghetti
mess of code!