Fixing export and missing subcommands
Both of these subcommands rely on the global index. Loading this
global index and acting upon it while tests are running is
problematic, however. Walking that index results in being unable to
load files that are being created and deleted in other tests.

Therefore, these tests get their own minimal index that only indexes
the Markdown files in the original checkout: README.md, index.md, and
the Markdown files in the the "themes" tree.

This also means that missingCli and exportCli must accept a pointer to
an index and use that instead of the global index.

Add a test for language in the export feed.

Add a test for a JSON feed.

Both of these tests used to fail because the exportCli funtion did not
copy p.Body to it.Body so when the template calls .Languages the
result was always empty.